Rename legacy handlers
