+++ cmd/courier/main.go
@@ -24,8 +24,7 @@ import (
 	_ "github.com/nyaruka/courier/handlers/discord"
 	_ "github.com/nyaruka/courier/handlers/dmark"
 	_ "github.com/nyaruka/courier/handlers/external"
-	_ "github.com/nyaruka/courier/handlers/facebook"
-	_ "github.com/nyaruka/courier/handlers/facebookapp"
+	_ "github.com/nyaruka/courier/handlers/facebook_legacy"
 	_ "github.com/nyaruka/courier/handlers/firebase"
 	_ "github.com/nyaruka/courier/handlers/freshchat"
 	_ "github.com/nyaruka/courier/handlers/globe"
@@ -45,6 +44,7 @@ import (
 	_ "github.com/nyaruka/courier/handlers/mblox"
 	_ "github.com/nyaruka/courier/handlers/messagebird"
 	_ "github.com/nyaruka/courier/handlers/messangi"
+	_ "github.com/nyaruka/courier/handlers/meta"
 	_ "github.com/nyaruka/courier/handlers/mtarget"
 	_ "github.com/nyaruka/courier/handlers/mtn"
 	_ "github.com/nyaruka/courier/handlers/nexmo"
@@ -66,7 +66,7 @@ import (
 	_ "github.com/nyaruka/courier/handlers/vk"
 	_ "github.com/nyaruka/courier/handlers/wavy"
 	_ "github.com/nyaruka/courier/handlers/wechat"
-	_ "github.com/nyaruka/courier/handlers/whatsapp"
+	_ "github.com/nyaruka/courier/handlers/whatsapp_legacy"
 	_ "github.com/nyaruka/courier/handlers/yo"
 	_ "github.com/nyaruka/courier/handlers/zenvia"
