+++ handlers/meta/meta.go
@@ -1,4 +1,4 @@
-package facebookapp
+package meta
 
 import (
 	"bytes"
