+++ handlers/whatsapp_legacy/whatsapp_test.go
@@ -1,4 +1,4 @@
-package whatsapp
+package whatsapp_legacy
 
 import (
 	"context"
